export/backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
python-dotenv==1.0.1

# Database
psycopg2-binary==2.9.11

# LangChain
langchain==0.3.0
langchain-core==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0

# Vector Search
faiss-cpu==1.8.0
numpy==1.26.0

# Document Processing
pypdf==4.0.0
python-docx==1.1.0
python-pptx==0.6.21
openpyxl==3.1.2
pandas==2.1.0
unstructured==0.10.0
markdown==3.5.0

# Utilities
tqdm==4.66.0